requirements.txt
Pillow
pillow-heif
pytest
PyYAML
